Add community feedback page with AI filtering
